Aparece a resposta para o usuario

+++ server/src/main/java/com/server/controller/QuestionController.java
@@ -16,7 +16,7 @@ public class QuestionController {
     	questionDao.add(question);
     }
     
-    public Question find(Integer questionId) {
+    public Question find(Long questionId) {
     	return questionDao.find(questionId);
     }
